Blood_Group_Detection/requirements.txt
numpy==1.26.4
opencv-python==4.9.0.80
matplotlib==3.8.3
Pillow==10.3.0
